move orient dependent class into orient specific package
