+++ nexus-repository-conan/src/test/java/org/sonatype/repository/conan/upgrade/ConanUpgrade_1_1_Test.groovy
@@ -1,15 +1,12 @@
 /*
  * Sonatype Nexus (TM) Open Source Version
  * Copyright (c) 2017-present Sonatype, Inc.
- * All rights reserved. Includes the third-party code listed at http://links.sonatype
- * .com/products/nexus/oss/attributions.
+ * All rights reserved. Includes the third-party code listed at http://links.sonatype.com/products/nexus/oss/attributions.
  *
- * This program and the accompanying materials are made available under the terms of the Eclipse Public License
- * Version 1.0,
+ * This program and the accompanying materials are made available under the terms of the Eclipse Public License Version 1.0,
  * which accompanies this distribution and is available at http://www.eclipse.org/legal/epl-v10.html.
  *
- * Sonatype Nexus (TM) Professional Version is available from Sonatype, Inc. "Sonatype" and "Sonatype Nexus" are
- * trademarks
+ * Sonatype Nexus (TM) Professional Version is available from Sonatype, Inc. "Sonatype" and "Sonatype Nexus" are trademarks
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
@@ -21,8 +18,8 @@ import org.sonatype.nexus.orient.OIndexNameBuilder
 import org.sonatype.nexus.orient.testsupport.DatabaseInstanceRule
 import org.sonatype.repository.conan.internal.AssetKind
 import org.sonatype.repository.conan.internal.ConanFormat
-import org.sonatype.repository.conan.internal.hosted.ConanHostedRecipe
-import org.sonatype.repository.conan.internal.proxy.v1.ConanProxyRecipe
+import org.sonatype.repository.conan.internal.orient.hosted.ConanHostedRecipe
+import org.sonatype.repository.conan.internal.orient.proxy.v1.ConanProxyRecipe
 
 import com.orientechnologies.orient.core.collate.OCaseInsensitiveCollate
 import com.orientechnologies.orient.core.db.record.OIdentifiable
