+++ nexus-repository-conan/src/test/resources/org/sonatype/repository/conan/internal/orient/metadata/conanmanifest.txt
